embedding-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
sentence-transformers==2.7.0
python-dotenv==1.0.1
pytest==8.3.2
